Use non-naive datetime for Python 2.7 and move utils to module_utils.compat.datetime
